feat(post): 实现分类的移动操作

+++ src/main/java/cn/caldm/www/post_context/infrastructure/repository/BlogPostCategoryRelationRepositoryImpl.java
@@ -12,7 +12,9 @@ import org.springframework.stereotype.Repository;
 import org.springframework.util.CollectionUtils;
 
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  *
@@ -88,6 +90,41 @@ public class BlogPostCategoryRelationRepositoryImpl implements BlogPostCategoryR
         LambdaQueryWrapper<BlogPostCategoryRelationPO> wrapper = new LambdaQueryWrapper<>();
         wrapper.eq(BlogPostCategoryRelationPO::getPostId, postId)
                 .in(BlogPostCategoryRelationPO::getCategoryId, categoryIds);
+        relationMapper.delete(wrapper);
+    }
+
+    @Override
+    public List<BlogPostCategoryRelation> findByCategoryIdsAndIsDirect(Collection<Long> categoryIds, Boolean isDirect) {
+        if (CollectionUtils.isEmpty(categoryIds)) {
+            return Collections.emptyList();
+        }
+        LambdaQueryWrapper<BlogPostCategoryRelationPO> wrapper = new LambdaQueryWrapper<>();
+        wrapper.in(BlogPostCategoryRelationPO::getCategoryId, categoryIds)
+                .eq(BlogPostCategoryRelationPO::getIsDirect, isDirect);
+        List<BlogPostCategoryRelationPO> pos = relationMapper.selectList(wrapper);
+        return pos.stream().map(relationAssembler::toDomain).collect(Collectors.toList());
+    }
+
+    @Override
+    public List<BlogPostCategoryRelation> findByPostIdsAndIsDirect(Collection<Long> postIds, Boolean isDirect) {
+        if (CollectionUtils.isEmpty(postIds)) {
+            return Collections.emptyList();
+        }
+        LambdaQueryWrapper<BlogPostCategoryRelationPO> wrapper = new LambdaQueryWrapper<>();
+        wrapper.in(BlogPostCategoryRelationPO::getPostId, postIds)
+                .eq(BlogPostCategoryRelationPO::getIsDirect, isDirect);
+        List<BlogPostCategoryRelationPO> pos = relationMapper.selectList(wrapper);
+        return pos.stream().map(relationAssembler::toDomain).collect(Collectors.toList());
+    }
+
+    @Override
+    public void deleteByPostIdsAndIsDirect(Collection<Long> postIds, Boolean isDirect) {
+        if (CollectionUtils.isEmpty(postIds)) {
+            return;
+        }
+        LambdaQueryWrapper<BlogPostCategoryRelationPO> wrapper = new LambdaQueryWrapper<>();
+        wrapper.in(BlogPostCategoryRelationPO::getPostId, postIds)
+                .eq(BlogPostCategoryRelationPO::getIsDirect, isDirect);
         relationMapper.delete(wrapper);
     }
 }
